finished 1.4, created pthread in light_sampler.c

+++ circular_buffer.h
@@ -9,11 +9,11 @@ typedef struct circular_buffer
     void* buffer_end;
     int length;
     int sz;
-    int count; // # of items in the buffer
+    int count; // # of samples scanned (could be bigger than buffer size)
     bool full;
 } circular_buffer;
 
-void buffer_init(circular_buffer *cb, int bufferSize, int sz);
+void buffer_init(circular_buffer *cb, int bufferSize, size_t sz);
 void buffer_free(circular_buffer *cb);
 void buffer_AddData(circular_buffer *cb, void* item);
